// src/app/shared/models/insurance.interface.ts
export interface IInsurance {
    car_value_and_sum_insured_market_value_rid: number;
    maker: string;
    model: string;
    year_model: string;
    description: string;
    new_price: number;
    from_date: string;
    to_date: string;
    market_value: number;
    min_sum_insure: number;
    max_sum_insure: number;
    screen_checked_flag: boolean;
    brand_rid: number;
    model_rid: number;
}